Feat: divide Layer component of IcePlate
